rename controller to handler
